// src/app/core/models/enums/apiErrorCode.ts
export enum ApiErrorCode {
  None = 0,
  ValidationErrors = 1,
  UnKnown = 2,
  NotFound = 3,
  CannotDelete = 4,
  InvalidOTP = 5,
  ExistMobile = 6,
  Unauthorize = 7,
  CannotSend = 8,
  NotActive = 9,
  CannotEdit = 10,
  UnauthorizeTokenIsBlackListed = 11,
  ExistEmail = 12,
  ManagementNotFound = 13,
  MediaNotFound = 14,
  ExistNationalNumber = 15,
  MobileOrPasswordNotCorrect = 16,
  NoAccountForEmail = 17,
  NotVerified = 18,
}